♻️ refactor: 직무 관련 DTO 및 직렬화 로직 네이밍 정리 및 타입 개선
- JobCategoryDto를 ReviewJobCategoryDto와 MemberJobCategoryDto로 분리
- String 기반 group/detail 필드를 JobGroup/JobDetail 타입으로 변경
- 관련 도메인, 엔드포인트, 테스트 클래스에 해당 변경점 적용
