+++ src/main/kotlin/picklab/backend/review/entrypoint/request/ReviewJobCategoryDto.kt
@@ -5,7 +5,7 @@ import jakarta.validation.constraints.NotNull
 import picklab.backend.job.domain.enums.JobDetail
 import picklab.backend.job.domain.enums.JobGroup
 
-data class JobCategoryDto(
+data class ReviewJobCategoryDto(
     @field:NotNull(message = "직무 대분류는 필수 입력값입니다.")
     @field:Schema(description = "관심 직무 대분류", example = "DEVELOPMENT")
     val jobGroup: JobGroup,
